requirements.txt
python-telegram-bot>=13.7,<14.0
Pillow==8.3.0
PyHyphen==4.0.0
pytz
git+https://gitlab.com/HirschHeissIch/ptbstats.git@v1.3.1